move setAttribute to render utils
