+++ core/src/plugins/render/directives/attr/attr.ts
@@ -2,7 +2,7 @@ import { isElementNode } from "@radish/htmlcrunch";
 import { assert, assertExists } from "@std/assert";
 import { Handler, handlerFor } from "@radish/effect-system";
 import { render } from "$effects/render.ts";
-import { setAttribute } from "../../common.ts";
+import { setAttribute } from "../../utils/setAttribute.ts";
 import { contextLookup } from "../../state.ts";
 
 export const handleAttrDirective = handlerFor(
